Move common parts into vrr app service
